+++ packages/atomic/src/components/search/atomic-sort-expression/atomic-sort-expression.stories.tsx
@@ -1,4 +1,4 @@
-import defaultStory from '../../../../.storybook/default-story';
+import defaultStory from 'atomic-storybook/default-story';
 
 const {defaultModuleExport, exportedStory} = defaultStory(
   'Atomic/SortDropdown/SortExpression',