+++ packages/atomic/src/components/search/result-template-components/atomic-result-image/atomic-result-image.stories.tsx
@@ -1,4 +1,4 @@
-import defaultResultComponentStory from '../../../../../.storybook/default-result-component-story';
+import defaultResultComponentStory from 'atomic-storybook/default-result-component-story';
 
 // TODO: This will require KIT-1178 to actually be usable properly
 const {defaultModuleExport, exportedStory} = defaultResultComponentStory(