+++ packages/quantic/force-app/main/default/lwc/quanticResultList/quanticResultList.js
@@ -141,8 +141,10 @@ export default class QuanticResultList extends LightningElement {
   }
 
   get fields() {
-    if (this.fieldsToInclude.trim() === '') return [];
-    return this.fieldsToInclude.split(',').map((field) => field.trim());
+    return this.fieldsToInclude
+      .split(',')
+      .map((field) => field.trim())
+      .filter((field) => field.length > 0);
   }
 
   get hasResults() {