+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.22.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.21.3...@coveo/atomic@2.22.0) (2023-04-03)
+
+### Features
+
+- **headless,atomic:** add organizationEndpoints configuration option for headless ([#2799](https://github.com/coveo/ui-kit/issues/2799)) ([c868d6d](https://github.com/coveo/ui-kit/commit/c868d6d529b51713318ae6839285610c6136f793))
+
+## [2.21.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.21.2...@coveo/atomic@2.21.3) (2023-04-03)
+
+### Bug Fixes
+
+- hide footer when empty ([#2791](https://github.com/coveo/ui-kit/issues/2791)) ([e7718e0](https://github.com/coveo/ui-kit/commit/e7718e0fd81f094afdc652cbf14955773ded8869))
+
+## [2.21.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.21.1...@coveo/atomic@2.21.2) (2023-03-23)
+
+### Bug Fixes
+
+- **atomic:** made smart snippet source title re-render when changed ([#2795](https://github.com/coveo/ui-kit/issues/2795) ([c98c490](https://github.com/coveo/ui-kit/commit/c98c490fbc129e80866d615cabd6c9c6bd445c0e))
+
+## [2.21.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.21.0...@coveo/atomic@2.21.1) (2023-03-22)
+
+**Note:** Version bump only for package @coveo/atomic
+
+# [2.21.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.20.2...@coveo/atomic@2.21.0) (2023-03-20)
+
+### Bug Fixes
+
+- **atomic:** multiple accessibility fixes ([#2782](https://github.com/coveo/ui-kit/issues/2782)) ([d90d353](https://github.com/coveo/ui-kit/commit/d90d3530a63e46f6e7f2cbf829c1cfa400398de8))
+
+### Features
+
+- recommendations error component ([#2787](https://github.com/coveo/ui-kit/issues/2787)) ([2d88e87](https://github.com/coveo/ui-kit/commit/2d88e8743970a4585fac807986420bc2c90b89ea))
+
 ## [2.20.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.20.1...@coveo/atomic@2.20.2) (2023-03-15)
 
 ### Bug Fixes